Implemented new data model for Entropy metric.

+++ base/coding/metrics/lib/src/main/java/com/puresol/coding/metrics/halstead/HalsteadMetric.java
@@ -168,7 +168,7 @@ public class HalsteadMetric extends CodeRangeEvaluator {
 		return result.getImplementationTime();
 	}
 
-	public double getProgramLength() {
+	public int getProgramLength() {
 		return result.getProgramLength();
 	}
 
@@ -184,7 +184,7 @@ public class HalsteadMetric extends CodeRangeEvaluator {
 		return result.getTotalOperators();
 	}
 
-	public double getVocabularySize() {
+	public int getVocabularySize() {
 		return result.getVocabularySize();
 	}
